requirements.txt
absl-py==1.2.0
astunparse==1.6.3
beautifulsoup4==4.11.1
bert-for-tf2==0.14.9
boto3==1.23.10
botocore==1.26.10
bs4==0.0.1
cachetools==5.2.0
certifi==2022.5.18.1
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.3
cryptography==37.0.2
cycler==0.11.0
distlib==0.3.4
filelock==3.7.0
Flask==2.1.2
Flask-Cors==3.0.10
flatbuffers==1.12
fonttools==4.34.4
fuzzy-c-means==1.6.4
gast==0.4.0
gensim==4.2.0
google-auth==2.9.1
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.47.0
h11==0.13.0
h5py==3.7.0
huggingface-hub==0.9.1
idna==3.3
importlib-metadata==4.11.4
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.0
joblib==1.1.0
keras==2.9.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
libclang==14.0.6
lxml==4.9.1
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.5.2
moto==3.1.11
networkx==2.8.6
nltk==3.7
numpy==1.23.1
oauthlib==3.2.0
opt-einsum==3.3.0
packaging==21.3
pandas==1.4.3
params-flow==0.8.2
Pillow==9.2.0
platformdirs==2.5.2
protobuf==4.21.7
py-params==0.10.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydantic==1.9.2
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.1
PyYAML==6.0
regex==2022.7.25
requests==2.27.1
requests-oauthlib==1.3.1
responses==0.21.0
rsa==4.9
s3transfer==0.5.2
scikit-fuzzy==0.4.2
scikit-learn==1.1.2
scipy==1.9.0
sentence-transformers==2.2.2
sentencepiece==0.1.97
six==1.16.0
sklearn==0.0
smart-open==6.2.0
soupsieve==2.3.2.post1
tabulate==0.8.10
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.9.1
tensorflow-estimator==2.9.0
tensorflow-hub==0.12.0
tensorflow-io-gcs-filesystem==0.26.0
termcolor==1.1.0
threadpoolctl==3.1.0
tokenizers==0.12.1
torch==1.12.1
torchvision==0.13.1
tqdm==4.64.0
transformers==4.21.2
typer==0.4.2
typing_extensions==4.3.0
urllib3==1.26.9
virtualenv==20.14.1
Werkzeug==2.1.2
wrapt==1.14.1
xmltodict==0.13.0
zipp==3.8.0
